requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiortc==1.6.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
audioread==3.0.1
authlib==1.6.0
av==11.0.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.16
email-validator==2.2.0
exceptiongroup==1.3.0
fastmcp==2.10.1
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-crc32c==1.7.1
google-genai==1.24.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h264-profile-level-id==1.0.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ifaddr==0.2.0
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
mcp==1.10.1
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
numba==0.61.2
numpy==2.2.6
openai==1.93.0
openapi-pydantic==0.5.1
packaging==25.0
pillow==10.4.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyee==11.1.0
pygments==2.19.2
pylibsrtp==0.12.0
pyopenssl==25.1.0
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
referencing==0.36.2
requests==2.31.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
sdp-transform==1.1.0
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==2.3.6
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.47.1
tenacity==8.5.0
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
videosdk==0.2.0
videosdk-agents==0.0.18
videosdk-plugins-google==0.0.18
videosdk-plugins-simli==0.0.18
vsaiortc==0.0.8
websockets==15.0.1
yarl==1.20.1
